move ByteUnit to package of only usage
